noscope: add no-scope game mode
The no-scope gamemode is enabled via `sm_noscope`. In this mode, each
player spawns with only a sniper rifle regardless of their chosen class.
If the player attempts to shoot the sniper while scoped, the weapon is
dropped and the player is warned. Damage from a scoped sniper rifle is
also prevented.